test: extract abstract testing classes
