+++ src/elements/abstract-intl-provider-element.spec.ts
@@ -1,70 +1,16 @@
 import {describe, it, expect} from '@jest/globals';
-import {nothing} from 'lit';
-import {property} from 'lit/decorators.js';
 
-import {createTestPage} from '../testing';
-import AbstractIntlProviderElement from './abstract-intl-provider-element';
+import {
+  TestIntlProviderElement,
+  createTestPage,
+} from '../testing';
 import HTMLIntlLocaleElement from './locale/locale';
 
-class FakeIntlApi {
-  static supportedLocalesOf(list: string | string[]) {
-    const supportedLocales = ['ar', 'en', 'es', 'ja', 'fr', 'zh', 'zh-Hant'];
-    if (list.includes('veryveryinvalid')) {
-      throw new RangeError();
-    }
-    if (Array.isArray(list)) {
-      return list.filter(locale => supportedLocales.includes(locale));
-    } else if (typeof list === 'string') {
-      return supportedLocales.includes(list) ? [list] : [];
-    }
-    return [];
-  }
-
-  format(unit: string): string {
-    switch (unit) {
-      case 'day':
-        return 'day';
-      case 'year':
-        return 'year';
-      default:
-        return '';
-    }
-  }
-}
-
-class TestIntlElement extends AbstractIntlProviderElement {
-  // @ts-ignore
-  protected static override intlApi = FakeIntlApi;
-
-  protected static override consumerElementNames = new Set(['bar']);
-
-  #intlObject!: FakeIntlApi;
-
-  // @ts-ignore
-  get intlObject(): FakeIntlApi {
-    return this.#intlObject;
-  }
-
-  @property({attribute: 'format-unit'})
-  // @ts-ignore
-  formatUnit = 'day';
-
-  resolvedOptions(): any {
-    return {};
-  }
-
-  override render() {
-    this.#intlObject = new FakeIntlApi();
-
-    return nothing;
-  }
-}
-
-customElements.define('intl-foo', TestIntlElement);
+customElements.define('intl-foo', TestIntlProviderElement);
 
 describe('AbstractIntlProviderElement', () => {
   it('handles multiple locales and ignores unsupported and invalid ones', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: 'intl-foo',
       html: `
         <intl-foo locales="he ja veryveryveryinvalid">
@@ -77,7 +23,7 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   it('trims `locales` attribute before parsing', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: 'intl-foo',
       html: `
         <intl-foo locales="   en      ja  ">
@@ -90,7 +36,7 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   it('should add `none` role if role is missing', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: 'intl-foo',
       html: `
         <intl-foo></intl-foo>
@@ -102,7 +48,7 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   it('should not override author defined role', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: 'intl-foo',
       html: `
         <intl-foo role="option"></intl-foo>
@@ -114,7 +60,7 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   it('rejects invalid atttibute values', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: 'intl-foo',
       html: `
         <intl-foo locales="en" format-unit="year"></intl-foo>
@@ -131,7 +77,7 @@ describe('AbstractIntlProviderElement', () => {
 
   describe('locale list determination prioritization', () => {
     test('1. `locales` attribute', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale tag="zh">
@@ -150,7 +96,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     it('keeps locale list based on `locales` attribute when ancestor changed', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <div lang="es">
@@ -168,7 +114,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     test('2. `lang` attribute', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale tag="zh">
@@ -187,7 +133,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     test('3. `locales-from` attribute', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale tag="zh">
@@ -205,7 +151,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     test('4.1. closest ancestor’s `lang`', async () => {
-      await createTestPage<TestIntlElement>({
+      await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale tag="zh">
@@ -216,14 +162,14 @@ describe('AbstractIntlProviderElement', () => {
           </intl-locale>
         `,
       });
-      const els = document.querySelectorAll('intl-foo') as NodeListOf<TestIntlElement>;
+      const els = document.querySelectorAll('intl-foo') as NodeListOf<TestIntlProviderElement>;
 
       expect(els[0]!.localeList.value).toBe('es');
       expect(els[1]!.localeList.value).toBe('es');
     });
 
     test('4.2. closest ancestor `intl-locale` element', async () => {
-      await createTestPage<TestIntlElement>({
+      await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <div lang="es">
@@ -236,7 +182,7 @@ describe('AbstractIntlProviderElement', () => {
           </div>
         `,
       });
-      const els = document.querySelectorAll('intl-foo') as NodeListOf<TestIntlElement>;
+      const els = document.querySelectorAll('intl-foo') as NodeListOf<TestIntlProviderElement>;
 
       expect(els[0]!.localeList.value).toBe('zh');
       expect(els[1]!.localeList.value).toBe('zh');
@@ -244,7 +190,7 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   it('observes `locales` attribute changes', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: 'intl-foo',
       html: `
         <intl-foo locales="en"></intl-foo>
@@ -260,7 +206,7 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   it('only updates `locales` if it has been defined by user', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: 'intl-foo',
       html: `
         <intl-foo lang="en"></intl-foo>
@@ -286,7 +232,7 @@ describe('AbstractIntlProviderElement', () => {
 
   describe('observes `locales` attribute removal and fallback', () => {
     test('1.1. added `lang` attribute', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: 'intl-foo',
         html: `
           <intl-foo locales="ar"></intl-foo>
@@ -303,7 +249,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     test('1.2. existing `lang` attribute', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: 'intl-foo',
         html: `
           <intl-foo locales="ar" lang="zh"></intl-foo>
@@ -319,7 +265,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     test('2. `locales-from` attribute', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale tag="zh">
@@ -339,7 +285,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     test('3.1. closest ancestor’s `lang`', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: 'intl-foo',
         html: `
           <div lang="es">
@@ -357,7 +303,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     test('3.2. closest ancestor `intl-locale` element', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale tag="zh">
@@ -376,7 +322,7 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   it('observes `lang` attribute changes', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: 'intl-foo',
       html: `
         <intl-foo lang="en"></intl-foo>
@@ -393,7 +339,7 @@ describe('AbstractIntlProviderElement', () => {
 
   describe('observes `lang` attribute removal and fallback', () => {
     test('1. `locales-from` attribute', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale tag="zh">
@@ -413,7 +359,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     test('2.1. closest ancestor’s `lang`', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: 'intl-foo',
         html: `
           <div lang="es">
@@ -431,7 +377,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     test('2.2. closest ancestor `intl-locale` element', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale tag="zh">
@@ -450,7 +396,7 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   it('uses locales from `intl-locale` elements associated by `locales-from`', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: ['intl-foo', 'intl-locale'],
       html: `
         <intl-locale id="my-ja" tag="ja"></intl-locale>
@@ -469,7 +415,7 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   it('observes the `locales-from` attribute and updates the locale list', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: ['intl-foo', 'intl-locale'],
       html: `
         <intl-locale id="my-ja" tag="ja"></intl-locale>
@@ -487,7 +433,7 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   it('observes the `locales-from` attribute removal and updates the locale list', async () => {
-    const page = await createTestPage<TestIntlElement>({
+    const page = await createTestPage<TestIntlProviderElement>({
       element: ['intl-foo', 'intl-locale'],
       html: `
         <div lang="es">
@@ -508,7 +454,7 @@ describe('AbstractIntlProviderElement', () => {
 
   describe('observes `intl-locale` elements associated by `locales-from` and updates the locale list', () => {
     it('observes `intl-locale`’s locale changes', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale id="my-ja" tag="ja"></intl-locale>
@@ -530,7 +476,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     it('observes `intl-locale` removal', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale id="my-ja" tag="ja"></intl-locale>
@@ -558,7 +504,7 @@ describe('AbstractIntlProviderElement', () => {
     });
 
     it('observes `intl-locale` removal and fallback to ancestor', async () => {
-      const page = await createTestPage<TestIntlElement>({
+      const page = await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <div lang="es">
@@ -584,11 +530,11 @@ describe('AbstractIntlProviderElement', () => {
   });
 
   describe('observes ancestors for `lang` attribute and `intl-locale` element changes', () => {
-    let el1: TestIntlElement;
-    let el2: TestIntlElement;
+    let el1: TestIntlProviderElement;
+    let el2: TestIntlProviderElement;
 
     beforeEach(async () => {
-      await createTestPage<TestIntlElement>({
+      await createTestPage<TestIntlProviderElement>({
         element: ['intl-foo', 'intl-locale'],
         html: `
           <intl-locale tag="zh">
@@ -599,8 +545,8 @@ describe('AbstractIntlProviderElement', () => {
           </intl-locale>
         `,
       });
-      el1 = document.getElementById('el1')! as TestIntlElement;
-      el2 = document.getElementById('el2')! as TestIntlElement;
+      el1 = document.getElementById('el1')! as TestIntlProviderElement;
+      el2 = document.getElementById('el2')! as TestIntlProviderElement;
 
       expect(el1.localeList.value).toBe('es');
       expect(el2.localeList.value).toBe('zh');
