+++ src/testing/test-intl-provider-element.ts
@@ -1,0 +1,34 @@
+import {nothing} from 'lit';
+import {property} from 'lit/decorators.js';
+
+import AbstractIntlProviderElement from '../elements/abstract-intl-provider-element';
+import {FakeIntlApi} from './fake-intl-api';
+
+export default class TestIntlProviderElement extends AbstractIntlProviderElement {
+  // @ts-ignore
+  protected static override intlApi = FakeIntlApi;
+
+  protected static override consumerElementNames = new Set(['bar']);
+
+  #intlObject!: FakeIntlApi;
+
+  // @ts-ignore
+  get intlObject(): FakeIntlApi {
+    return this.#intlObject;
+  }
+
+  @property({attribute: 'format-unit'})
+  // @ts-ignore
+  formatUnit = 'day';
+
+  resolvedOptions(): any {
+    return {};
+  }
+
+  override render() {
+    this.#intlObject = new FakeIntlApi();
+
+    return nothing;
+  }
+}
+
